Renamed Gitlab::Merge to Gitlab::Satellite::MergeAction
